+++ editor/include/adminer.inc.php
@@ -7,7 +7,7 @@ class Adminer {
 	/** @var array<string, string[]|string> */ private array $values = array(); // [table => options or one description]
 
 	function name(): string {
-		return "<a href='https://www.adminer.org/editor/'" . target_blank() . " id='h1'><img src='../adminer/static/logo.png' width='24' height='24' alt='' id='logo'>" . lang('Editor') . "</a>";
+		return "<a href='https://www.adminer.org/editor/'" . target_blank() . " id='h1'><img src='" . DIR . "static/logo.png' width='24' height='24' alt='' id='logo'>" . lang('Editor') . "</a>";
 	}
 
 	//! driver, ns